requirements.txt
# langchain==0.1.20
# langchain-google-genai==1.0.3
# langgraph==0.0.69
# langchain-community==0.0.38
# python-dotenv==1.0.0
# streamlit==1.32.0

langchain==0.2.0
langgraph==0.2.20
langchain-community==0.2.0
langchain-google-genai==1.0.5
python-dotenv==1.0.0
streamlit==1.32.0
